+++ src/ethereum/forks/amsterdam/fork.py
@@ -641,7 +641,6 @@ def make_receipt(
     tx: Transaction,
     error: Optional[EthereumException],
     cumulative_gas_used: Uint,
-    gas_spent: Uint,
     logs: Tuple[Log, ...],
 ) -> Bytes | Receipt:
     """
@@ -655,9 +654,7 @@ def make_receipt(
         Error in the top level frame of the transaction, if any.
     cumulative_gas_used :
         The total gas used so far in the block after the transaction was
-        executed. This is the gas used before refunds.
-    gas_spent :
-        The gas actually spent by this transaction after refunds.
+        executed. This is the gas used after refunds.
     logs :
         The logs produced by the transaction.
 
@@ -672,7 +669,6 @@ def make_receipt(
         cumulative_gas_used=cumulative_gas_used,
         bloom=logs_bloom(logs),
         logs=logs,
-        gas_spent=gas_spent,
     )
 
     return encode_receipt(tx, receipt)
@@ -1104,14 +1100,14 @@ def process_transaction(
     ):
         destroy_account(tx_state, block_env.coinbase)
 
+    block_output.cumulative_gas_used += tx_gas_used
     block_output.block_gas_used += block_gas_used_in_tx
     block_output.blob_gas_used += tx_blob_gas_used
 
     receipt = make_receipt(
         tx,
         tx_output.error,
-        block_output.block_gas_used,
-        tx_gas_used,
+        block_output.cumulative_gas_used,
         all_logs,
     )