+++ src/ethereum_spec_tools/evm_tools/t8n/t8n_types.py
@@ -449,8 +449,6 @@ class Result:
                 }
                 logs_json.append(log_dict)
             receipt_dict["logs"] = logs_json
-            if hasattr(receipt, "gas_spent"):
-                receipt_dict["gasSpent"] = hex(receipt.gas_spent)
 
             receipts_json.append(receipt_dict)